Maintenances of Forms
- active and inactive forms
-some bug fix in research.edit

+++ app/Http/Controllers/AcademicDevelopment/CollegeDepartmentAwardController.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\Controller;
 use App\Models\CollegeDepartmentAward;
 use Illuminate\Support\Facades\Storage;
 use App\Models\CollegeDepartmentAwardDocument;
+use App\Models\FormBuilder\AcademicDevelopmentForm;
 
 class CollegeDepartmentAwardController extends Controller
 {
@@ -30,6 +31,8 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function create()
     {
+        if(AcademicDevelopmentForm::where('id', 6)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $awardFields = DB::select("CALL get_academic_development_fields_by_form_id(6)");
 
         return view('academic-development.college-department-award.create', compact('awardFields'));
@@ -43,6 +46,8 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function store(Request $request)
     {
+        if(AcademicDevelopmentForm::where('id', 6)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document']);
 
         $college_department_award = CollegeDepartmentAward::create($input);
@@ -82,6 +87,8 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function show(CollegeDepartmentAward $college_department_award)
     {
+        if(AcademicDevelopmentForm::where('id', 6)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $awardFields = DB::select("CALL get_academic_development_fields_by_form_id(6)");
 
         $documents = CollegeDepartmentAwardDocument::where('college_department_award_id', $college_department_award->id)->get()->toArray();
@@ -99,6 +106,8 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function edit(CollegeDepartmentAward $college_department_award)
     {
+        if(AcademicDevelopmentForm::where('id', 6)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $awardFields = DB::select("CALL get_academic_development_fields_by_form_id(6)");
 
         $documents = CollegeDepartmentAwardDocument::where('college_department_award_id', $college_department_award->id)->get()->toArray();
@@ -117,6 +126,8 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function update(Request $request, CollegeDepartmentAward $college_department_award)
     {
+        if(AcademicDevelopmentForm::where('id', 6)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document']);
 
         $college_department_award->update($input);
@@ -155,12 +166,16 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function destroy(CollegeDepartmentAward $college_department_award)
     {
+        if(AcademicDevelopmentForm::where('id', 6)->pluck('is_active')->first() == 0)
+            return view('inactive');
         CollegeDepartmentAwardDocument::where('college_department_award_id', $college_department_award->id)->delete();
         $college_department_award->delete();
         return redirect()->route('college-department-award.index')->with('award_success', 'Your Accomplishment in Awards and Recognition Received by the College and Department has been deleted.');
     }
 
     public function removeDoc($filename){
+        if(AcademicDevelopmentForm::where('id', 6)->pluck('is_active')->first() == 0)
+            return view('inactive');
         CollegeDepartmentAwardDocument::where('filename', $filename)->delete();
         return true;
     }
