+++ app/Http/Controllers/AcademicDevelopment/ReferenceController.php
@@ -2,15 +2,16 @@
 
 namespace App\Http\Controllers\AcademicDevelopment;
 
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
 use App\Models\Reference;
-use App\Models\ReferenceDocument;
+use Illuminate\Http\Request;
 use App\Models\TemporaryFile;
-use Illuminate\Support\Facades\Storage;
+use App\Models\ReferenceDocument;
 use Illuminate\Support\Facades\DB;
 use App\Models\Maintenance\College;
+use App\Http\Controllers\Controller;
 use App\Models\Maintenance\Department;
+use Illuminate\Support\Facades\Storage;
+use App\Models\FormBuilder\AcademicDevelopmentForm;
 
 class ReferenceController extends Controller
 {
@@ -36,6 +37,8 @@ class ReferenceController extends Controller
      */
     public function create()
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $referenceFields = DB::select("CALL get_academic_development_fields_by_form_id(1)");
 
         $colleges = College::all();
@@ -50,6 +53,8 @@ class ReferenceController extends Controller
      */
     public function store(Request $request)
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
 
         $rtmmi = Reference::create($input);
@@ -94,6 +99,8 @@ class ReferenceController extends Controller
      */
     public function show(Reference $rtmmi)
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $referenceDocuments = ReferenceDocument::where('reference_id', $rtmmi->id)->get()->toArray();
 
         $category = DB::select("CALL get_dropdown_name_by_id(".$rtmmi->category.")");
@@ -115,6 +122,8 @@ class ReferenceController extends Controller
      */
     public function edit(Reference $rtmmi)
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $referenceFields = DB::select("CALL get_academic_development_fields_by_form_id(1)");
 
         $referenceDocuments = ReferenceDocument::where('reference_id', $rtmmi->id)->get()->toArray();
@@ -142,6 +151,8 @@ class ReferenceController extends Controller
      */
     public function update(Request $request, Reference $rtmmi)
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
 
         $rtmmi->update($input);
@@ -186,6 +197,8 @@ class ReferenceController extends Controller
      */
     public function destroy(Reference $rtmmi)
     {
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $rtmmi->delete();
         ReferenceDocument::where('reference_id', $rtmmi->id)->delete();
 
@@ -199,6 +212,8 @@ class ReferenceController extends Controller
     }
 
     public function removeDoc($filename){
+        if(AcademicDevelopmentForm::where('id', 1)->pluck('is_active')->first() == 0)
+            return view('inactive');
         ReferenceDocument::where('filename', $filename)->delete();
         Storage::delete('documents/'.$filename);
         return true;
