+++ app/Http/Controllers/AcademicDevelopment/SyllabusController.php
@@ -2,15 +2,16 @@
 
 namespace App\Http\Controllers\AcademicDevelopment;
 
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
 use App\Models\Syllabus;
-use Illuminate\Support\Facades\DB;
-use App\Models\SyllabusDocument;
+use Illuminate\Http\Request;
 use App\Models\TemporaryFile;
-use Illuminate\Support\Facades\Storage;
+use App\Models\SyllabusDocument;
+use Illuminate\Support\Facades\DB;
 use App\Models\Maintenance\College;
+use App\Http\Controllers\Controller;
 use App\Models\Maintenance\Department;
+use Illuminate\Support\Facades\Storage;
+use App\Models\FormBuilder\AcademicDevelopmentForm;
 
 class SyllabusController extends Controller
 {
@@ -36,6 +37,8 @@ class SyllabusController extends Controller
      */
     public function create()
     {
+        if(AcademicDevelopmentForm::where('id', 2)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $syllabusFields = DB::select("CALL get_academic_development_fields_by_form_id(2)");
 
         $colleges = College::all();
@@ -50,6 +53,8 @@ class SyllabusController extends Controller
      */
     public function store(Request $request)
     {
+        if(AcademicDevelopmentForm::where('id', 2)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
 
         $syllabus = Syllabus::create($input);
@@ -89,6 +94,8 @@ class SyllabusController extends Controller
      */
     public function show(Syllabus $syllabu)
     {
+        if(AcademicDevelopmentForm::where('id', 2)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $syllabusDocuments = SyllabusDocument::where('syllabus_id', $syllabu->id)->get()->toArray();
 
         $collegeAndDepartment = DB::select("CALL get_college_and_department_by_department_id(".$syllabu->department_id.")");
@@ -106,6 +113,8 @@ class SyllabusController extends Controller
      */
     public function edit(Syllabus $syllabu)
     {
+        if(AcademicDevelopmentForm::where('id', 2)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $syllabusFields = DB::select("CALL get_academic_development_fields_by_form_id(2)");
 
         $syllabusDocuments = SyllabusDocument::where('syllabus_id', $syllabu->id)->get()->toArray();
@@ -131,6 +140,8 @@ class SyllabusController extends Controller
      */
     public function update(Request $request, Syllabus $syllabu)
     {
+        if(AcademicDevelopmentForm::where('id', 2)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
 
         $syllabu->update($input);
@@ -170,6 +181,8 @@ class SyllabusController extends Controller
      */
     public function destroy(Syllabus $syllabu)
     {
+        if(AcademicDevelopmentForm::where('id', 2)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $syllabu->delete();
         SyllabusDocument::where('syllabus_id', $syllabu->id)->delete();
 
@@ -178,6 +191,8 @@ class SyllabusController extends Controller
     }
 
     public function removeDoc($filename){
+        if(AcademicDevelopmentForm::where('id', 2)->pluck('is_active')->first() == 0)
+            return view('inactive');
         SyllabusDocument::where('filename', $filename)->delete();
         Storage::delete('documents/'.$filename);
         return true;
