+++ app/Http/Controllers/ExtensionPrograms/PartnershipController.php
@@ -11,6 +11,7 @@ use App\Models\PartnershipDocument;
 use App\Http\Controllers\Controller;
 use App\Models\Maintenance\Department;
 use Illuminate\Support\Facades\Storage;
+use App\Models\FormBuilder\ExtensionProgramForm;
 use App\Models\FormBuilder\ExtensionProgramField;
 
 class PartnershipController extends Controller
@@ -33,6 +34,8 @@ class PartnershipController extends Controller
      */
     public function create()
     {
+        if(ExtensionProgramForm::where('id', 5)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $partnershipFields = DB::select("CALL get_extension_program_fields_by_form_id('5')");
 
         $colleges = College::all();
@@ -47,6 +50,8 @@ class PartnershipController extends Controller
      */
     public function store(Request $request)
     {
+        if(ExtensionProgramForm::where('id', 5)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
 
         $partnership = Partnership::create($input);
@@ -87,6 +92,8 @@ class PartnershipController extends Controller
      */
     public function show(Partnership $partnership)
     {
+        if(ExtensionProgramForm::where('id', 5)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $partnershipFields = DB::select("CALL get_extension_program_fields_by_form_id('5')");
 
         $documents = PartnershipDocument::where('partnership_id', $partnership->id)->get()->toArray();
@@ -104,6 +111,8 @@ class PartnershipController extends Controller
      */
     public function edit(Partnership $partnership)
     {
+        if(ExtensionProgramForm::where('id', 5)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $partnershipFields = DB::select("CALL get_extension_program_fields_by_form_id('5')");
 
         $collegeAndDepartment = Department::join('colleges', 'colleges.id', 'departments.college_id')
@@ -130,6 +139,8 @@ class PartnershipController extends Controller
      */
     public function update(Request $request, Partnership $partnership)
     {
+        if(ExtensionProgramForm::where('id', 5)->pluck('is_active')->first() == 0)
+            return view('inactive');
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
 
         $partnership->update($input);
@@ -169,12 +180,16 @@ class PartnershipController extends Controller
      */
     public function destroy(Partnership $partnership)
     {
+        if(ExtensionProgramForm::where('id', 5)->pluck('is_active')->first() == 0)
+            return view('inactive');
         PartnershipDocument::where('partnership_id', $partnership->id)->delete();
         $partnership->delete();
         return redirect()->route('partnership.index')->with('partnership_success', 'Your accomplishment in Partnership/ Linkages/ Network has been deleted.');
     }
 
     public function removeDoc($filename){
+        if(ExtensionProgramForm::where('id', 5)->pluck('is_active')->first() == 0)
+            return view('inactive');
         PartnershipDocument::where('filename', $filename)->delete();
         return true;
     }
